+++ pkg.nix
@@ -1,12 +1,12 @@
-{ mkDerivation, base, bytestring, exceptions, network
-, safe-exceptions, stdenv, transformers
+{ mkDerivation, base, bytestring, network, safe-exceptions, socks
+, stdenv, transformers
 }:
 mkDerivation {
   pname = "network-simple";
-  version = "0.4.2";
+  version = "0.4.3";
   src = ./.;
   libraryHaskellDepends = [
-    base bytestring exceptions network safe-exceptions transformers
+    base bytestring network safe-exceptions socks transformers
   ];
   homepage = "https://github.com/k0001/network-simple";
   description = "Simple network sockets usage patterns";